Renamed Schema to MapResolver
