+++ lib/Extension/ClassMover/Rpc/ClassMoveHandler.php
@@ -2,7 +2,7 @@
 
 namespace Phpactor\Extension\ClassMover\Rpc;
 
-use Phpactor\Container\Schema;
+use Phpactor\MapResolver\Resolver;
 use Phpactor\Extension\Rpc\Response\OpenFileResponse;
 use Phpactor\Extension\Rpc\Response\Input\TextInput;
 use Phpactor\Extension\ClassMover\Application\Logger\NullLogger;
@@ -40,7 +40,7 @@ class ClassMoveHandler extends AbstractHandler
         return 'move_class';
     }
 
-    public function configure(Schema $schema): void
+    public function configure(Resolver $schema): void
     {
         $schema->setDefaults([
             self::PARAM_SOURCE_PATH => null,
