+++ lib/Extension/CodeTransform/Rpc/OverrideMethodHandler.php
@@ -2,7 +2,7 @@
 
 namespace Phpactor\Extension\CodeTransform\Rpc;
 
-use Phpactor\Container\Schema;
+use Phpactor\MapResolver\Resolver;
 use Phpactor\WorseReflection\Reflector;
 use InvalidArgumentException;
 use Phpactor\Extension\Rpc\Response\Input\ListInput;
@@ -46,7 +46,7 @@ class OverrideMethodHandler extends AbstractHandler
         return self::NAME;
     }
 
-    public function configure(Schema $schema): void
+    public function configure(Resolver $schema): void
     {
         $schema->setDefaults([
             self::PARAM_PATH => null,
