+++ lib/Extension/Rpc/Handler.php
@@ -2,13 +2,13 @@
 
 namespace Phpactor\Extension\Rpc;
 
-use Phpactor\Container\Schema;
+use Phpactor\MapResolver\Resolver;
 
 interface Handler
 {
     public function name(): string;
 
-    public function configure(Schema $schema): void;
+    public function configure(Resolver $schema): void;
 
     public function handle(array $arguments);
 }
