+++ tests/Integration/ApplicationTest.php
@@ -7,7 +7,7 @@ use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Output\BufferedOutput;
 use Phpactor\Tests\IntegrationTestCase;
 use Phpactor\Bootstrap;
-use Phpactor\Container\InvalidConfig;
+use Phpactor\MapResolver\InvalidConfig;
 
 class ApplicationTest extends IntegrationTestCase
 {
